requirements.txt
flask
Flask-Mail
Flask-SQLAlchemy
Jinja2